Refactor reporters
